Finish writing the calculateBouncedNoiseLevel function.
This function is based on the calculateDirectNoiseLevel; it will draw a cone of 30deg with the soundSource at the top and the sink at the middle of the base. Fifty bounce points (unique) are made, and the final sound level is calculated calculated from the sum of the bounced noise.

Several mathutils regarding array manipulation, vector calculation, and equally spaced points generation is written.

A barebone constructor for a dummy BaseUnit object is written.